Implement performTransaction and getTransaction
Test informally

+++ src/main/kotlin/it/polito/group19/lab2/dto/WalletDto.kt
@@ -4,7 +4,7 @@ import it.polito.group19.lab2.domain.Customer
 
 data class WalletDto(
     var wid: Long,
-    var balance: Double,
+    var balance: Float,
     var customerId: Long,
     var rechargeIds: MutableList<Long>,
     var purchaseIds: MutableList<Long>,
